gui/FaceForge/requirements.txt
mediapipe>=0.10.0
Pillow>=9.0.0
numpy>=1.21.0
opencv-python==4.10.0.84
opencv-contrib-python==4.10.0.84